+++ marbelle/backend/marbelle/settings/dev.py
@@ -3,6 +3,7 @@ Development settings for marbelle project.
 """
 
 import os
+
 from .base import *  # noqa: F403,F405
 
 # SECURITY WARNING: don't run with debug turned on in production!